Rewrite subscribeBy support
Switching from including the rxkotlin jars to using a stub fixes an issue that prevented resolving the rxkotlin methods in tests, which lead to a sub-optimal implementation of the check.

The onError check has been simplified by checking if the named parameter is supplied in the subscribeBy call, which should fix all false positives that previously happened.

Fixes #31